chore: refactor to fix test error
