+++ src/pages/api/bot-firewall/get-blocked-ips.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { BlockedIp, BlockedIpDbModel } from '../../../server/botd-firewall/blockedIpsDatabase';
+import { getBlockedIps } from '../../../server/botd-firewall/blockedIpsDatabase';
 
-export default async function handler(req: NextApiRequest, res: NextApiResponse<BlockedIp[]>) {
-  const blockedIps = await BlockedIpDbModel.findAll();
+export default async function handler(req: NextApiRequest, res: NextApiResponse<string[]>) {
+  const blockedIps = await getBlockedIps();
   res.status(200).json(blockedIps);
 }
