+++ src/server/botd-firewall/blockedIpsDatabase.ts
@@ -1,5 +1,6 @@
 import { Attributes, DataTypes, InferAttributes, InferCreationAttributes, Model } from 'sequelize';
 import { sequelize } from '../server';
+import { MAX_BLOCKED_IPS } from './updateFirewallRule';
 
 interface BlockedIpAttributes
   extends Model<InferAttributes<BlockedIpAttributes>, InferCreationAttributes<BlockedIpAttributes>> {
@@ -19,6 +20,14 @@ export const BlockedIpDbModel = sequelize.define<BlockedIpAttributes>('blocked_i
 
 BlockedIpDbModel.sync({ force: false });
 export type BlockedIp = Attributes<BlockedIpAttributes>;
+
+export const getBlockedIps = async (): Promise<string[]> => {
+  const blockedIps = await BlockedIpDbModel.findAll({
+    order: [['timestamp', 'DESC']],
+    limit: MAX_BLOCKED_IPS,
+  });
+  return blockedIps.map((ip) => ip.ip);
+};
 
 export const saveBlockedIp = async (ip: string) => {
   BlockedIpDbModel.upsert({
